[fix] allow exporting custom Metal kernels without a Metal backend
- Move metal_kernel graph construction to backend/common so it builds on
  all platforms
- During export tracing without Metal, record the kernel on a placeholder
  GPU stream which the importing process remaps to a real stream
- Outside of export tracing, calling the kernel without Metal still raises
  an error

Fixes #3240
