+++ docs/src/dev/custom_metal_kernels.rst
@@ -62,7 +62,7 @@ The full function signature will be generated using:
     so we add ``device float16_t* out``.
 * Template parameters passed using ``template``
     In the above, ``template=[("T", mx.float32)]`` adds a template of ``template <typename T>`` to the function
-    and instantiates the template with ``custom_kernel_myexp_float_float16_float16<float>``.
+    and instantiates the template with ``custom_kernel_myexp_float_float16_t_float16_t<float>``.
     Template parameters can be ``mx.core.Dtype``, ``int`` or ``bool``.
 * Metal attributes used in ``source`` such as ``[[thread_position_in_grid]]``
     These will be added as function arguments.
@@ -73,7 +73,7 @@ Putting this all together, the generated function signature for ``myexp`` is as 
 .. code-block:: cpp
 
   template <typename T>
-  [[kernel]] void custom_kernel_myexp_float_float16_float16(
+  [[kernel]] void custom_kernel_myexp_float_float16_t_float16_t(
     const device float16_t* inp [[buffer(0)]],
     device float16_t* out [[buffer(1)]],
     uint3 thread_position_in_grid [[thread_position_in_grid]]) {
@@ -84,7 +84,7 @@ Putting this all together, the generated function signature for ``myexp`` is as 
 
   }
 
-  template [[host_name("custom_kernel_myexp_float_float16_float16")]] [[kernel]] decltype(custom_kernel_myexp_float_float16_float16<float>) custom_kernel_myexp_float_float16_float16<float>;
+  template [[host_name("custom_kernel_myexp_float_float16_t_float16_t")]] [[kernel]] decltype(custom_kernel_myexp_float_float16_t_float16_t<float>) custom_kernel_myexp_float_float16_t_float16_t<float>;
 
 Note: ``grid`` and ``threadgroup`` are parameters to the Metal `dispatchThreads
 <https://developer.apple.com/documentation/metal/mtlcomputecommandencoder/2866532-dispatchthreads>`_
