+++ mlx/backend/metal/no_metal.cpp
@@ -3,7 +3,6 @@
 #include <stdexcept>
 
 #include "mlx/backend/metal/metal.h"
-#include "mlx/fast.h"
 
 namespace mlx::core {
 
@@ -23,20 +22,5 @@ device_info() {
 };
 
 } // namespace metal
-
-namespace fast {
-
-CustomKernelFunction metal_kernel(
-    const std::string&,
-    const std::vector<std::string>&,
-    const std::vector<std::string>&,
-    const std::string&,
-    const std::string&,
-    bool,
-    bool) {
-  throw std::runtime_error("[metal_kernel] No Metal back-end.");
-}
-
-} // namespace fast
 
 } // namespace mlx::core
